requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.1
colorama==0.4.6
Flask==3.1.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
python-telegram-bot==22.5
rsa==4.9.1
sniffio==1.3.1
Telethon==1.41.2
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.4
yarl==1.22.0
